Refactor `utils.ts` into `common` module
