+++ app/ide-desktop/.prettierrc.yaml
@@ -25,7 +25,9 @@ importOrder:
   - ""
   - "<THIRD_PARTY_MODULES>"
   - ""
-  - "^enso-.*$"
+  - "^enso-assets"
+  - ""
+  - "^enso-"
   - ""
   - "^[.].*/[^./]*$"
   - ""
