+++ app/ide-desktop/lib/client/paths.ts
@@ -1,6 +1,6 @@
 /** @file This module defines paths within the client distribution's resources. */
 
-import * as utils from '../../utils'
+import * as requireEnv from 'enso-common/requireEnv'
 
 // ==========================
 // === Paths to resources ===
@@ -12,25 +12,25 @@ export const PROJECT_MANAGER_BUNDLE = 'enso'
 
 /** Distribution directory for IDE. */
 export function getIdeDirectory(): string {
-    return utils.requireEnv('ENSO_BUILD_IDE')
+    return requireEnv.requireEnv('ENSO_BUILD_IDE')
 }
 
 /** Distribution directory for GUI. */
 export function getGuiDirectory(): string {
-    return utils.requireEnv('ENSO_BUILD_GUI')
+    return requireEnv.requireEnv('ENSO_BUILD_GUI')
 }
 
 /** Path to the project manager bundle root. */
 export function getProjectManagerBundlePath(): string {
-    return utils.requireEnv('ENSO_BUILD_PROJECT_MANAGER')
+    return requireEnv.requireEnv('ENSO_BUILD_PROJECT_MANAGER')
 }
 
 /** Path to the project manager executable relative to the PM bundle root. */
 export function getProjectManagerInBundlePath(): string {
-    return utils.requireEnv('ENSO_BUILD_PROJECT_MANAGER_IN_BUNDLE_PATH')
+    return requireEnv.requireEnv('ENSO_BUILD_PROJECT_MANAGER_IN_BUNDLE_PATH')
 }
 
 /** Version of the Engine (backend) that is bundled along with this client build. */
 export function getBundledEngineVersion(): string {
-    return utils.requireEnv('ENSO_BUILD_IDE_BUNDLED_ENGINE_VERSION')
+    return requireEnv.requireEnv('ENSO_BUILD_IDE_BUNDLED_ENGINE_VERSION')
 }
