+++ app/ide-desktop/lib/common/src/json.ts
@@ -1,0 +1,8 @@
+/** @file Constants related to JSON serialization and deserialization. */
+
+// =================
+// === Constants ===
+// =================
+
+/** Indent size for outputting JSON. */
+export const INDENT_SIZE = 4
