+++ app/ide-desktop/lib/content/esbuild-config.ts
@@ -20,7 +20,7 @@ import esbuildPluginCopyDirectories from 'esbuild-plugin-copy-directories'
 import esbuildPluginTime from 'esbuild-plugin-time'
 import esbuildPluginYaml from 'esbuild-plugin-yaml'
 
-import * as utils from '../../utils'
+import * as requireEnv from 'enso-common/requireEnv'
 
 import BUILD_INFO from '../../../../build.json' assert { type: 'json' }
 
@@ -58,9 +58,9 @@ export interface Arguments extends PassthroughArguments {
 
 /** Get arguments from the environment. */
 export function argumentsFromEnv(passthroughArguments: PassthroughArguments): Arguments {
-    const wasmArtifacts = utils.requireEnv('ENSO_BUILD_GUI_WASM_ARTIFACTS')
-    const assetsPath = utils.requireEnv('ENSO_BUILD_GUI_ASSETS')
-    const outputPath = pathModule.resolve(utils.requireEnv('ENSO_BUILD_GUI'), 'assets')
+    const wasmArtifacts = requireEnv.requireEnv('ENSO_BUILD_GUI_WASM_ARTIFACTS')
+    const assetsPath = requireEnv.requireEnv('ENSO_BUILD_GUI_ASSETS')
+    const outputPath = pathModule.resolve(requireEnv.requireEnv('ENSO_BUILD_GUI'), 'assets')
     return { ...passthroughArguments, wasmArtifacts, assetsPath, outputPath }
 }
