+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/components/keyboardShortcut.tsx
@@ -6,7 +6,8 @@ import CtrlKeyIcon from 'enso-assets/ctrl_key.svg'
 import OptionKeyIcon from 'enso-assets/option_key.svg'
 import ShiftKeyIcon from 'enso-assets/shift_key.svg'
 import WindowsKeyIcon from 'enso-assets/windows_key.svg'
-import * as detect from 'enso-common/src/detect'
+
+import * as detect from 'enso-common/detect'
 
 import SvgMask from '../../authentication/components/svgMask'
 import * as shortcutsProvider from '../../providers/shortcuts'
